refactor: moved KeepScreenOn to Utils.kt and merged redundant functions
